Moved domain classes to new domain package
