+++ src/com/movierentals/domain/MovieCategory.java
@@ -1,4 +1,4 @@
-package com.movierentals;
+package com.movierentals.domain;
 
 
 public enum MovieCategory {
